#10334: Migrate real_bw to TTNN
