+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -641,8 +641,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.complex_recip_bw
 
-.. autofunction:: tt_lib.tensor.real_bw
-
 .. autofunction:: tt_lib.tensor.complex_mul_bw
 
 .. autofunction:: tt_lib.tensor.complex_div_bw
