+++ docs/source/ttnn/ttnn/ttnn/real_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.real_bw:
+
+ttnn.real_bw
+###################
+
+.. autofunction:: ttnn.real_bw
