+++ tests/ttnn/unit_tests/operations/backward/complex_ops/test_backward_real.py
@@ -1,0 +1,73 @@
+# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+
+import math
+from pathlib import Path
+import sys
+
+import torch
+
+import tt_lib as ttl
+import pytest
+import ttnn
+from loguru import logger
+
+from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_pcc, comp_equal, comp_allclose
+
+from models.utility_functions import is_wormhole_b0
+from tests.ttnn.unit_tests.operations.backward.complex_ops.backward_complex_utility_funcs import (
+    Complex,
+    convert_to_torch_tensor,
+    random_complex_tensor,
+)
+
+
+@pytest.mark.parametrize(
+    "memcfg",
+    (
+        ttl.tensor.MemoryConfig(ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.DRAM),
+        ttl.tensor.MemoryConfig(ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.L1),
+    ),
+    ids=["out_DRAM", "out_L1"],
+)
+@pytest.mark.parametrize("dtype", ((ttl.tensor.DataType.BFLOAT16,)))
+@pytest.mark.parametrize("bs", ((1, 1), (1, 2), (2, 2)))
+@pytest.mark.parametrize("hw", ((32, 64), (320, 384)))
+def test_level2_real_bw(bs, hw, memcfg, dtype, device, function_level_defaults):
+    input_shape = torch.Size([bs[0], bs[1], hw[0], hw[1]])
+
+    in_data = random_complex_tensor(input_shape, (-90, 90), (-70, 70))
+    in_data.requires_grad = True
+
+    input_tensor = ttl.tensor.complex_tensor(
+        ttl.tensor.Tensor(in_data.real, dtype).to(ttl.tensor.Layout.TILE).to(device, memcfg),
+        ttl.tensor.Tensor(in_data.imag, dtype).to(ttl.tensor.Layout.TILE).to(device, memcfg),
+    )
+
+    grad_data = random_complex_tensor(input_shape, (-50, 50), (-60, 60))
+    grad_data = grad_data.real
+    grad_tensor = ttl.tensor.Tensor(
+        ttl.tensor.Tensor(grad_data, dtype).to(ttl.tensor.Layout.TILE).to(device, memcfg),
+    )
+    tt_dev = ttnn.real_bw(grad_tensor, input_tensor, memory_config=memcfg)
+    in_data.retain_grad()
+
+    tt_dev = convert_to_torch_tensor(tt_dev)
+
+    pyt_y = torch.real(in_data)
+
+    pyt_y.backward(gradient=grad_data)
+
+    grad_res_real = torch.real(in_data.grad)
+    grad_res_imag = torch.imag(in_data.grad)
+
+    tt_cpu = [torch.cat((grad_res_real, grad_res_imag), dim=-1)]
+
+    for i in range(len(tt_dev)):
+        if is_wormhole_b0():
+            passing, output = comp_pcc(tt_cpu[i], tt_dev[i])
+        else:
+            passing, output = comp_pcc(tt_cpu[i], tt_dev[i])
+        logger.info(output)
+        assert passing
