+++ ttnn/cpp/ttnn/operations/eltwise/complex_unary_backward/device/complex_unary_backward_op.hpp
@@ -16,6 +16,7 @@ constexpr uint8_t DefaultQueueId = 0;
 enum class ComplexUnaryBackwardOpType {
     POLAR_BW,
     IMAG_BW,
+    REAL_BW,
 };
 
 //OpHandler_complex : get_function_complex
@@ -23,6 +24,7 @@ std::vector<ComplexTensor> _polar_bw(const ComplexTensor& grad, const ComplexTen
 
 //OpHandler_tensor_complex : get_function_tensor_complex
 std::vector<ComplexTensor> _imag_bw(const Tensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config);
+std::vector<ComplexTensor> _real_bw(const Tensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config);
 
 template <ComplexUnaryBackwardOpType OpType>
 struct OpHandler_complex;
@@ -41,6 +43,13 @@ template <>
 struct OpHandler_tensor_complex<ComplexUnaryBackwardOpType::IMAG_BW> {
     static std::vector<ComplexTensor> handle( const Tensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config ) {
         return _imag_bw(grad, input, output_mem_config);
+    }
+};
+
+template <>
+struct OpHandler_tensor_complex<ComplexUnaryBackwardOpType::REAL_BW> {
+    static std::vector<ComplexTensor> handle( const Tensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config ) {
+        return _real_bw(grad, input, output_mem_config);
     }
 };
